Updated avoid_unnecessary_return_variable rule and its visitor to use AnalysisRule.
