+++ lib/src/lints/avoid_unnecessary_return_variable/visitors/avoid_unnecessary_return_variable_visitor.dart
@@ -1,79 +1,81 @@
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/ast/visitor.dart';
 import 'package:analyzer/dart/element/element.dart';
-import 'package:collection/collection.dart';
+import 'package:solid_lints/src/common/visitors/select_expression_identifiers_visitor.dart';
+import 'package:solid_lints/src/lints/avoid_unnecessary_return_variable/avoid_unnecessary_return_variable_rule.dart';
+import 'package:solid_lints/src/lints/avoid_unnecessary_return_variable/visitors/return_variable_usage_visitor.dart';
 
-/// This visitor is searches all uses of a single local variable,
-/// including variable declaration.
+/// Visitor for [AvoidUnnecessaryReturnVariableRule].
 class AvoidUnnecessaryReturnVariableVisitor extends RecursiveAstVisitor<void> {
-  /// The problem expects that exactly 1 mention of return variable.
-  /// VariableDeclarationStatement doesn't count when visiting SimpleIdentifier.
-  /// Any other amount of variable mentions implies that it is used somewhere
-  /// except return, so its existence is justified.
-  static const _badStatementCount = 1;
-
-  final LocalVariableElement _returnVariableElement;
-
-  bool _foundTokensBetweenDeclarationAndReturn = false;
-  VariableDeclaration? _variableDeclaration;
-  int _variableStatementCounter = 0;
-
-  final ReturnStatement _returnStatement;
-
-  /// After visiting holds info about whether there are any tokens
-  /// between variable declaration and return statement
-  bool get foundTokensBetweenDeclarationAndReturn =>
-      _foundTokensBetweenDeclarationAndReturn;
-
-  /// Returns statement of local variable declaration
-  VariableDeclaration? get variableDeclaration => _variableDeclaration;
+  /// The rule associated with this visitor.
+  final AvoidUnnecessaryReturnVariableRule rule;
 
   /// Creates a new instance of [AvoidUnnecessaryReturnVariableVisitor].
-  AvoidUnnecessaryReturnVariableVisitor(
-    this._returnVariableElement,
-    this._returnStatement,
-  );
-
-  /// Defines whether the variables is used in return statement only.
-  bool hasBadStatementCount() =>
-      _variableStatementCounter == _badStatementCount;
+  AvoidUnnecessaryReturnVariableVisitor(this.rule);
 
   @override
-  void visitVariableDeclarationStatement(VariableDeclarationStatement node) {
-    if (_collectVariableDeclaration(node)) {
-      _checkTokensInBetween(node, _returnStatement);
-    }
-    super.visitVariableDeclarationStatement(node);
-  }
-
-  @override
-  void visitSimpleIdentifier(SimpleIdentifier node) {
-    if (node.element?.id == _returnVariableElement.id) {
-      _variableStatementCounter++;
+  void visitReturnStatement(ReturnStatement node) {
+    final expr = node.expression;
+    if (expr is! SimpleIdentifier) {
+      return;
     }
 
-    super.visitSimpleIdentifier(node);
-  }
-
-  bool _collectVariableDeclaration(VariableDeclarationStatement node) {
-    final targetVariable = node.variables.variables.firstWhereOrNull(
-      (v) => v.declaredFragment?.element.id == _returnVariableElement.id,
-    );
-    if (targetVariable == null) return false;
-
-    _variableDeclaration = targetVariable;
-    return true;
-  }
-
-  void _checkTokensInBetween(
-    VariableDeclarationStatement variableDeclaration,
-    ReturnStatement returnStatement,
-  ) {
-    final tokenBeforeReturn =
-        _returnStatement.findPrevious(_returnStatement.beginToken);
-
-    if (tokenBeforeReturn != variableDeclaration.endToken) {
-      _foundTokensBetweenDeclarationAndReturn = true;
+    final element = expr.element;
+    if (element is! LocalVariableElement) {
+      return;
     }
+
+    if (!element.isFinal && !element.isConst) {
+      return;
+    }
+
+    final block = node.parent;
+    if (block == null) return;
+
+    final returnVariableUsageVisitor =
+        ReturnVariableUsageVisitor(node, element);
+
+    block.visitChildren(returnVariableUsageVisitor);
+    if (!returnVariableUsageVisitor.hasBadStatementCount()) return;
+
+    //it is 100% bad if return statement follows declaration
+    if (!returnVariableUsageVisitor.foundTokensBetweenDeclarationAndReturn) {
+      rule.reportAtNode(node);
+      return;
+    }
+
+    final declaration = returnVariableUsageVisitor.variableDeclaration;
+
+    //check if immutable
+    final initializer = declaration?.initializer;
+    if (initializer == null) return;
+
+    if (!_isExpressionImmutable(initializer)) return;
+
+    rule.reportAtNode(node);
+
+    super.visitReturnStatement(node);
+  }
+
+  bool _isExpressionImmutable(Expression expr) {
+    final visitor = SelectExpressionIdentifiersVisitor();
+    expr.accept(visitor);
+
+    return visitor.identifiers.every(_isSimpleIdentifierImmutable);
+  }
+
+  bool _isSimpleIdentifierImmutable(SimpleIdentifier identifier) {
+    switch (identifier.element) {
+      case final VariableElement variable:
+        return variable.isFinal || variable.isConst;
+
+      case ClassElement _:
+        return true;
+
+      case GetterElement(:final PropertyInducingElement variable):
+        return variable.isFinal || variable.isConst;
+    }
+
+    return false;
   }
 }
